% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_a_centrigrados.R
\name{fahrenheit_a_centigrados}
\alias{fahrenheit_a_centigrados}
\title{Coversion de la temperatura en Fahrenheit a Centigrados}
\usage{
fahrenheit_a_centigrados(temp_fahrenheit)
}
\arguments{
\item{temp_fahrenheit}{un vector numérico de la temperatura en Fahrenheit}
}
\value{
Un vector numérico de la temperatura en grados centígrados.
}
\description{
La función \code{fahrenheit_a_centigrados()} toma un valor numérico de la temperatura en Fahrenheit y calcula la equivalencia en grados centígrados.
}
\details{
Estos son detalles extras de la función.
}
\examples{
fahrenheit_a_centigrados(100)

}
